NotificationOperationをBright.NotificationOperationsに移動
